Removed Xunit1 dependencies from Akka.TestKit.Tests and Akka.FSharp.Tests projects
