+++ src/core/Akka.TestKit.Tests/TestEventListenerTests/AllTestForEventFilterBase.cs
@@ -1,30 +1,33 @@
 ﻿//-----------------------------------------------------------------------
 // <copyright file="AllTestForEventFilterBase.cs" company="Akka.NET Project">
-//     Copyright (C) 2009-2015 Lightbend Inc. <http://www.lightbend.com>
-//     Copyright (C) 2013-2015 Akka.NET project <https://github.com/akkadotnet/akka.net>
+//     Copyright (C) 2009-2016 Lightbend Inc. <http://www.lightbend.com>
+//     Copyright (C) 2013-2016 Akka.NET project <https://github.com/akkadotnet/akka.net>
 // </copyright>
 //-----------------------------------------------------------------------
 
 using System;
 using System.Threading.Tasks;
 using Akka.Event;
-using Akka.TestKit;
 using Xunit;
 using Xunit.Sdk;
 
-namespace Akka.Testkit.Tests.TestEventListenerTests
+namespace Akka.TestKit.Tests.Xunit2.TestEventListenerTests
 {
     public abstract class AllTestForEventFilterBase<TLogEvent> : EventFilterTestBase where TLogEvent : LogEvent
     {
         // ReSharper disable ConvertToLambdaExpression
+        private readonly EventFilterFactory _testingEventFilter;
 
         protected AllTestForEventFilterBase(string config)
             : base(config)
         {
             LogLevel = Logging.LogLevelFor<TLogEvent>();
+            // ReSharper disable once VirtualMemberCallInContructor
+            _testingEventFilter = CreateTestingEventFilter();
         }
 
         protected LogLevel LogLevel { get; private set; }
+        protected abstract EventFilterFactory CreateTestingEventFilter();
 
         protected void LogMessage(string message)
         {
@@ -39,24 +42,24 @@ namespace Akka.Testkit.Tests.TestEventListenerTests
         protected abstract void PublishMessage(object message, string source);
 
         [Fact]
-        public void SingleMessageIsIntercepted()
+        public void Single_message_is_intercepted()
         {
-            EventFilter.ForLogLevel(LogLevel).ExpectOne(() => LogMessage("whatever"));
+            _testingEventFilter.ForLogLevel(LogLevel).ExpectOne(() => LogMessage("whatever"));
             TestSuccessful = true;
         }
 
 
         [Fact]
-        public void CanInterceptMessagesWhenStartIsSpecified()
+        public void Can_intercept_messages_when_start_is_specified()
         {
-            EventFilter.ForLogLevel(LogLevel, start: "what").ExpectOne(() => LogMessage("whatever"));
+            _testingEventFilter.ForLogLevel(LogLevel, start: "what").ExpectOne(() => LogMessage("whatever"));
             TestSuccessful = true;
         }
 
         [Fact]
-        public void DoNotInterceptMessagesWhenStartDoesNotMatch()
+        public void Do_not_intercept_messages_when_start_does_not_match()
         {
-            EventFilter.ForLogLevel(LogLevel, start: "what").ExpectOne(() =>
+            _testingEventFilter.ForLogLevel(LogLevel, start: "what").ExpectOne(() =>
             {
                 LogMessage("let-me-thru");
                 LogMessage("whatever");
@@ -66,14 +69,14 @@ namespace Akka.Testkit.Tests.TestEventListenerTests
         }
 
         [Fact]
-        public void CanInterceptMessagesWhenMessageIsSpecified()
+        public void Can_intercept_messages_when_message_is_specified()
         {
-            EventFilter.ForLogLevel(LogLevel, message: "whatever").ExpectOne(() => LogMessage("whatever"));
+            _testingEventFilter.ForLogLevel(LogLevel, message: "whatever").ExpectOne(() => LogMessage("whatever"));
             TestSuccessful = true;
         }
 
         [Fact]
-        public void DoNotInterceptMessagesWhenMessageDoesNotMatch()
+        public void Do_not_intercept_messages_when_message_does_not_match()
         {
             EventFilter.ForLogLevel(LogLevel, message: "whatever").ExpectOne(() =>
             {
@@ -85,16 +88,16 @@ namespace Akka.Testkit.Tests.TestEventListenerTests
         }
 
         [Fact]
-        public void CanInterceptMessagesWhenContainsIsSpecified()
+        public void Can_intercept_messages_when_contains_is_specified()
         {
-            EventFilter.ForLogLevel(LogLevel, contains: "ate").ExpectOne(() => LogMessage("whatever"));
+            _testingEventFilter.ForLogLevel(LogLevel, contains: "ate").ExpectOne(() => LogMessage("whatever"));
             TestSuccessful = true;
         }
 
         [Fact]
-        public void DoNotInterceptMessagesWhenContainsDoesNotMatch()
+        public void Do_not_intercept_messages_when_contains_does_not_match()
         {
-            EventFilter.ForLogLevel(LogLevel, contains: "eve").ExpectOne(() =>
+            _testingEventFilter.ForLogLevel(LogLevel, contains: "eve").ExpectOne(() =>
             {
                 LogMessage("let-me-thru");
                 LogMessage("whatever");
@@ -105,16 +108,16 @@ namespace Akka.Testkit.Tests.TestEventListenerTests
 
 
         [Fact]
-        public void CanInterceptMessagesWhenSourceIsSpecified()
+        public void Can_intercept_messages_when_source_is_specified()
         {
-            EventFilter.ForLogLevel(LogLevel, source: GetType().FullName).ExpectOne(() => LogMessage("whatever"));
+            _testingEventFilter.ForLogLevel(LogLevel, source: GetType().FullName).ExpectOne(() => LogMessage("whatever"));
             TestSuccessful = true;
         }
 
         [Fact]
-        public void DoNotInterceptMessagesWhenSourceDoesNotMatch()
+        public void Do_not_intercept_messages_when_source_does_not_match()
         {
-            EventFilter.ForLogLevel(LogLevel, source: "expected-source").ExpectOne(() =>
+            _testingEventFilter.ForLogLevel(LogLevel, source: "expected-source").ExpectOne(() =>
             {
                 PublishMessage("message", source: "expected-source");
                 PublishMessage("message", source: "let-me-thru");
@@ -126,9 +129,9 @@ namespace Akka.Testkit.Tests.TestEventListenerTests
 
 
         [Fact]
-        public void SpecifiedNumbersOfMessagesanBeIntercepted()
+        public void Specified_numbers_of_messagesan_be_intercepted()
         {
-            EventFilter.ForLogLevel(LogLevel).Expect(2, () =>
+            _testingEventFilter.ForLogLevel(LogLevel).Expect(2, () =>
             {
                 LogMessage("whatever");
                 LogMessage("whatever");
@@ -138,9 +141,9 @@ namespace Akka.Testkit.Tests.TestEventListenerTests
 
 
         [Fact]
-        public void MessagesCanBeMuted()
+        public void Messages_can_be_muted()
         {
-            EventFilter.ForLogLevel(LogLevel).Mute(() =>
+            _testingEventFilter.ForLogLevel(LogLevel).Mute(() =>
             {
                 LogMessage("whatever");
                 LogMessage("whatever");
@@ -150,9 +153,9 @@ namespace Akka.Testkit.Tests.TestEventListenerTests
 
 
         [Fact]
-        public void MessagesCanBeMutedFromNowOn()
+        public void Messages_can_be_muted_from_now_on()
         {
-            var unmutableFilter = EventFilter.ForLogLevel(LogLevel).Mute();
+            var unmutableFilter = _testingEventFilter.ForLogLevel(LogLevel).Mute();
             LogMessage("whatever");
             LogMessage("whatever");
             unmutableFilter.Unmute();
@@ -160,9 +163,9 @@ namespace Akka.Testkit.Tests.TestEventListenerTests
         }
 
         [Fact]
-        public void MessagesCanBeMutedFromNowOnWithUsing()
+        public void Messages_can_be_muted_from_now_on_with_using()
         {
-            using(EventFilter.ForLogLevel(LogLevel).Mute())
+            using(_testingEventFilter.ForLogLevel(LogLevel).Mute())
             {
                 LogMessage("whatever");
                 LogMessage("whatever");
@@ -172,18 +175,18 @@ namespace Akka.Testkit.Tests.TestEventListenerTests
 
 
         [Fact]
-        public void MakeSureAsyncWorks()
+        public void Make_sure_async_works()
         {
-            EventFilter.ForLogLevel(LogLevel).Expect(1, TimeSpan.FromMilliseconds(100), () =>
+            _testingEventFilter.ForLogLevel(LogLevel).Expect(1, TimeSpan.FromMilliseconds(100), () =>
             {
                 Task.Delay(TimeSpan.FromMilliseconds(10)).ContinueWith(t => { LogMessage("whatever"); });
             });
         }
 
         [Fact]
-        public void ChainManyFilters()
+        public void Chain_many_filters()
         {
-            EventFilter
+            _testingEventFilter
                 .ForLogLevel(LogLevel,message:"Message 1").And
                 .ForLogLevel(LogLevel,message:"Message 3")
                 .Expect(2,() =>
@@ -198,11 +201,11 @@ namespace Akka.Testkit.Tests.TestEventListenerTests
 
 
         [Fact]
-        public void ShouldTimeoutIfTooFewMessages()
+        public void Should_timeout_if_too_few_messages()
         {
-            var exception = XAssert.Throws<AssertException>(() =>
+            var exception = XAssert.Throws<TrueException>(() =>
             {
-                EventFilter.ForLogLevel(LogLevel).Expect(2, TimeSpan.FromMilliseconds(50), () =>
+                _testingEventFilter.ForLogLevel(LogLevel).Expect(2, TimeSpan.FromMilliseconds(50), () =>
                 {
                     LogMessage("whatever");
                 });
@@ -211,7 +214,7 @@ namespace Akka.Testkit.Tests.TestEventListenerTests
         }
 
         [Fact]
-        public void ShouldLogWhenNotMuting()
+        public void Should_log_when_not_muting()
         {
             const string message = "This should end up in the log since it's not filtered";
             LogMessage(message);
