+++ src/core/Akka.TestKit.Tests/TestEventListenerTests/ConfigTests.cs
@@ -1,17 +1,16 @@
 ﻿//-----------------------------------------------------------------------
 // <copyright file="ConfigTests.cs" company="Akka.NET Project">
-//     Copyright (C) 2009-2015 Lightbend Inc. <http://www.lightbend.com>
-//     Copyright (C) 2013-2015 Akka.NET project <https://github.com/akkadotnet/akka.net>
+//     Copyright (C) 2009-2016 Lightbend Inc. <http://www.lightbend.com>
+//     Copyright (C) 2013-2016 Akka.NET project <https://github.com/akkadotnet/akka.net>
 // </copyright>
 //-----------------------------------------------------------------------
 
 using System.Linq;
-using Akka.TestKit;
 using Xunit;
 
-namespace Akka.Testkit.Tests.TestEventListenerTests
+namespace Akka.TestKit.Tests.Xunit2.TestEventListenerTests
 {
-    public class ConfigTests : TestKit.Xunit.TestKit
+    public class ConfigTests : TestKit.Xunit2.TestKit
     {
         [Fact]
         public void TestEventListener_is_in_config_by_default()
