+++ src/core/Akka.TestKit.Tests/TestKitBaseTests/RemainingTests.cs
@@ -1,16 +1,16 @@
 ﻿//-----------------------------------------------------------------------
 // <copyright file="RemainingTests.cs" company="Akka.NET Project">
-//     Copyright (C) 2009-2015 Lightbend Inc. <http://www.lightbend.com>
-//     Copyright (C) 2013-2015 Akka.NET project <https://github.com/akkadotnet/akka.net>
+//     Copyright (C) 2009-2016 Lightbend Inc. <http://www.lightbend.com>
+//     Copyright (C) 2013-2016 Akka.NET project <https://github.com/akkadotnet/akka.net>
 // </copyright>
 //-----------------------------------------------------------------------
 
 using System;
 using Xunit;
 
-namespace Akka.Testkit.Tests.TestKitBaseTests
+namespace Akka.TestKit.Tests.Xunit2.TestKitBaseTests
 {
-    public class RemainingTests : Akka.TestKit.Xunit.TestKit
+    public class RemainingTests : TestKit.Xunit2.TestKit
     {
         [Fact]
         public void Throw_if_remaining_is_called_outside_Within()
